Create new file and copy paste tests
